requirements.txt
chainlit==1.0.200
python-dotenv==1.0.0
requests==2.31.0 